Write values back to wolf from other sources
Atm we can only change values of the wolf device through the iobroker GUI.
But if we have, lets say, mqtt enabled, and subscribed to a topic (to change a values with mqtt commands), the value was changed in the iobroker GUI but not send to the wolf device. With this change, we always also update the wolf device if the change is not coming from the wolf device itself.